refactored route_service.dart (split for single and multi);
