+++ src/flutter_app/lib/components/info/route_info.dart
@@ -1,4 +1,3 @@
-import 'package:climbing_diary/services/route_service.dart';
 import 'package:flutter/material.dart';
 import 'package:internet_connection_checker/internet_connection_checker.dart';
 
@@ -6,6 +5,7 @@ import '../../interfaces/ascent/ascent.dart';
 import '../../interfaces/ascent/ascent_style.dart';
 import '../../interfaces/ascent/ascent_type.dart';
 import '../../interfaces/multi_pitch_route/multi_pitch_route.dart';
+import '../../services/multi_pitch_route_service.dart';
 import '../my_text_styles.dart';
 
 class RouteInfo extends StatefulWidget {
@@ -19,7 +19,7 @@ class RouteInfo extends StatefulWidget {
 }
 
 class _RouteInfoState extends State<RouteInfo>{
-  RouteService routeService = RouteService();
+  final MultiPitchRouteService multiPitchRouteService = MultiPitchRouteService();
   bool online = false;
 
   void checkConnection() async {
@@ -38,7 +38,7 @@ class _RouteInfoState extends State<RouteInfo>{
   @override
   Widget build(BuildContext context) {
     return FutureBuilder<Ascent?>(
-      future: routeService.getBestAscent(widget.route, online),
+      future: multiPitchRouteService.getBestAscent(widget.route, online),
       builder: (context, snapshot) {
         if (snapshot.hasData) {
           Ascent? ascent = snapshot.data;
