+++ src/flutter_app/lib/pages/list_page/multi_pitch_route_details.dart
@@ -11,8 +11,8 @@ import '../../components/info/multi_pitch_route_info.dart';
 import '../../components/my_button_styles.dart';
 import '../../interfaces/multi_pitch_route/multi_pitch_route.dart';
 import '../../services/media_service.dart';
+import '../../services/multi_pitch_route_service.dart';
 import '../../services/pitch_service.dart';
-import '../../services/route_service.dart';
 
 class MultiPitchRouteDetails extends StatefulWidget {
   const MultiPitchRouteDetails({super.key, required this.route, required this.onNetworkChange});
@@ -26,7 +26,7 @@ class MultiPitchRouteDetails extends StatefulWidget {
 
 class _MultiPitchRouteDetailsState extends State<MultiPitchRouteDetails>{
   final MediaService mediaService = MediaService();
-  final RouteService routeService = RouteService();
+  final MultiPitchRouteService multiPitchRouteService = MultiPitchRouteService();
   final PitchService pitchService = PitchService();
 
   Future<List<String>> fetchURLs() {
@@ -46,7 +46,7 @@ class _MultiPitchRouteDetailsState extends State<MultiPitchRouteDetails>{
         var mediaId = await mediaService.uploadMedia(img);
         MultiPitchRoute multiPitchRoute = widget.route;
         multiPitchRoute.mediaIds.add(mediaId);
-        routeService.editMultiPitchRoute(multiPitchRoute.toUpdateMultiPitchRoute());
+        multiPitchRouteService.editMultiPitchRoute(multiPitchRoute.toUpdateMultiPitchRoute());
       }
     } else {
       List<XFile> images = await picker.pickMultiImage();
@@ -54,7 +54,7 @@ class _MultiPitchRouteDetailsState extends State<MultiPitchRouteDetails>{
         var mediaId = await mediaService.uploadMedia(img);
         MultiPitchRoute multiPitchRoute = widget.route;
         multiPitchRoute.mediaIds.add(mediaId);
-        routeService.editMultiPitchRoute(multiPitchRoute.toUpdateMultiPitchRoute());
+        multiPitchRouteService.editMultiPitchRoute(multiPitchRoute.toUpdateMultiPitchRoute());
       }
     }
     setState(() {});
@@ -90,7 +90,7 @@ class _MultiPitchRouteDetailsState extends State<MultiPitchRouteDetails>{
 
     void deleteImageCallback(String mediumId) {
       widget.route.mediaIds.remove(mediumId);
-      routeService.editMultiPitchRoute(UpdateMultiPitchRoute(
+      multiPitchRouteService.editMultiPitchRoute(UpdateMultiPitchRoute(
         id: widget.route.id,
         mediaIds: widget.route.mediaIds
       ));
